+++ packages/apptive_grid_form/lib/translation/l10n/translation_en.dart
@@ -38,4 +38,10 @@ class ApptiveGridLocalizedTranslation extends ApptiveGridTranslation {
   String get searchLocation => "Search location";
   @override
   String get searchLocationNoResult => "No results";
+  @override
+  String get attachmentSourceGallery => "Choose existing";
+  @override
+  String get attachmentSourceCamera => "Take Photo or Video";
+  @override
+  String get attachmentSourceFiles => "Select files";
 }