+++ packages/apptive_grid_form/test/common.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
 import 'package:http/http.dart' as http;
+import 'package:image_picker_platform_interface/image_picker_platform_interface.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:permission_handler_platform_interface/permission_handler_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -18,6 +19,12 @@ class MockApptiveGridCache extends Mock implements ApptiveGridCache {}
 class MockFilePicker extends Mock
     with MockPlatformInterfaceMixin
     implements FilePicker {}
+
+class MockImagePicker extends Mock
+    with MockPlatformInterfaceMixin
+    implements ImagePickerPlatform {}
+
+class MockXFile extends Mock implements XFile {}
 
 class MockGeolocator extends Mock
     with MockPlatformInterfaceMixin